feat: Initial assume-role CLI
This matches the behaviour of a hokey CoffeeScript script - either an
AWS profile can be given or a role ARN (and optionally a source profile
and/or external ID). When a profile is given, and it contains an
"mfa_serial", an MFA token will be read from stdin.

In either case, the output will be formatted for evaluation by a bash-
compatible shell.